+++ core/Cargo.toml
@@ -22,12 +22,12 @@ libc = "0.2.98"
 log = "0.4.14"
 parking_lot = "0.11.1"
 pin-project = "1.0.7"
+rustls = { version = "0.19.1", features = ["dangerous_configuration"] }
 rusty_v8 = "0.25.3"
 serde = { version = "1.0.126", features = ["derive"] }
 serde_json = { version = "1.0.64", features = ["preserve_order"] }
 serde_v8 = { version = "0.8.0" }
 url = { version = "2.2.2", features = ["serde"] }
-rustls = { version = "0.19.1", features = ["dangerous_configuration"] }
 webpki = "0.21.4"
 
 [[example]]
