+++ extensions/fetch/Cargo.toml
@@ -22,8 +22,8 @@ deno_web = { version = "0.44.0", path = "../web" }
 http = "0.2.4"
 lazy_static = "1.4.0"
 reqwest = { version = "0.11.4", default-features = false, features = ["rustls-tls", "stream", "gzip", "brotli"] }
+rustls = { version = "0.19.1", features = ["dangerous_configuration"] }
 serde = { version = "1.0.126", features = ["derive"] }
 tokio = { version = "1.8.1", features = ["full"] }
 tokio-stream = "0.1.7"
 tokio-util = "0.6.7"
-rustls = { version = "0.19.1", features = ["dangerous_configuration"] }
