+++ runtime/build.rs
@@ -64,8 +64,7 @@ fn create_runtime_snapshot(snapshot_path: &Path, files: Vec<PathBuf>) {
       false, // No --unstable.
     ),
     deno_net::init::<deno_net::NoNetPermissions>(
-      None,
-      false, // No --unstable.
+      None, false, // No --unstable.
       None,
     ),
     deno_ffi::init::<deno_ffi::NoFfiPermissions>(false),
